requirements.txt
playwright
beautifulsoup4
python-dotenv
python-telegram-bot[job-queue]
